# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for a simple private inventory and sales app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Company {
  id: number;
  name: string;
  slug: string;
}

export interface Product {
  id: number;
  name: string;
  brand?: string;
  content?: string;
  cost: number;
  salePrice: number;
  stock: number;
  minimumStock: number;
  isDemo?: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  brand?: string;
  content?: string;
  /** @minimum 0 */
  cost: number;
  /** @minimum 0 */
  salePrice: number;
  /** @minimum 0 */
  initialStock: number;
  /** @minimum 0 */
  minimumStock?: number;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  brand?: string;
  content?: string;
  /** @minimum 0 */
  cost?: number;
  /** @minimum 0 */
  salePrice?: number;
  /** @minimum 0 */
  minimumStock?: number;
}

export interface InventoryInput {
  /** @minimum 1 */
  quantity: number;
  note?: string;
}

export interface SaleItemInput {
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface SaleInput {
  /** @minItems 1 */
  items: SaleItemInput[];
}

export interface SaleItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  unitCost: number;
  subtotal: number;
}

export interface Sale {
  id: number;
  date: string;
  total: number;
  totalItems: number;
  estimatedProfit: number;
  items: SaleItem[];
}

export interface LowStockProduct {
  id: number;
  name: string;
  stock: number;
  minimumStock: number;
}

export interface DeudaMoisesInput {
  value: number;
}

export interface Dashboard {
  productCount: number;
  totalUnits: number;
  inventoryCostValue: number;
  inventorySaleValue: number;
  potentialProfit: number;
  todaySalesCount: number;
  todaySalesTotal: number;
  /** Valor monetario de la deuda de Moises David */
  deudaMoisesDavid?: number;
  /** Si el usuario actual puede editar la deuda */
  canEditDeudaMoises?: boolean;
  lowStockProducts: LowStockProduct[];
}

export interface InventoryReport {
  products: Product[];
  totalCostValue: number;
  totalSaleValue: number;
  potentialProfit: number;
}

export interface SalesReport {
  sales: Sale[];
  totalSold: number;
  saleCount: number;
  itemCount: number;
  estimatedProfit: number;
  /** @nullable */
  bestSellingProduct: string | null;
  lowStockProducts: LowStockProduct[];
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Not found
 */
export type NotFoundResponse = Error;

export type ListSalesParams = {
period?: ListSalesPeriod;
from?: string;
to?: string;
};

export type ListSalesPeriod = typeof ListSalesPeriod[keyof typeof ListSalesPeriod];


export const ListSalesPeriod = {
  today: 'today',
  last7: 'last7',
  thisMonth: 'thisMonth',
  previousMonth: 'previousMonth',
  all: 'all',
} as const;

export type GetInventoryReportParams = {
filter?: GetInventoryReportFilter;
};

export type GetInventoryReportFilter = typeof GetInventoryReportFilter[keyof typeof GetInventoryReportFilter];


export const GetInventoryReportFilter = {
  all: 'all',
  low: 'low',
  empty: 'empty',
} as const;

export type GetSalesReportParams = {
period?: GetSalesReportPeriod;
from?: string;
to?: string;
};

export type GetSalesReportPeriod = typeof GetSalesReportPeriod[keyof typeof GetSalesReportPeriod];


export const GetSalesReportPeriod = {
  today: 'today',
  last7: 'last7',
  thisMonth: 'thisMonth',
  previousMonth: 'previousMonth',
  all: 'all',
} as const;