# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for a simple private inventory and sales app
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  slug: string;
}